refactor(dashboard): rename useLocalModelStatus → useLocalModelStatuses; consume /local-models/status (Spec 2 SC38, SC39)
